VFB-315 refactor variable name and boolean logic
VFB-315 refactor default address setting logic and add new file for large function def
